backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langgraph
chromadb
pypdf
tavily-python
python-dotenv
langchain-google-genai
langchain-huggingface
sentence-transformers
pysqlite3-binary